Rename UInputMoney & UInputNumber components, small refactoring
