+++ src/types.ts
@@ -48,12 +48,12 @@ import UTextareaConfig from "./ui.form-textarea/config.ts";
 import ULabelConfig from "./ui.form-label/config.ts";
 import UColorPickerConfig from "./ui.other-theme-color-toggle/config.ts";
 import UInputConfig from "./ui.form-input/config.ts";
-import UInputNumberConfig from "./ui.form-input-number/config.ts";
+import UInputCounterConfig from "./ui.form-input-counter/config.ts";
 import UInputPasswordConfig from "./ui.form-input-password/config.ts";
 import UInputRatingConfig from "./ui.form-input-rating/config.ts";
 import UInputSearchConfig from "./ui.form-input-search/config.ts";
 import UInputFileConfig from "./ui.form-input-file/config.ts";
-import UInputMoneyConfig from "./ui.form-input-money/config.ts";
+import UInputNumberConfig from "./ui.form-input-number/config.ts";
 import UDataListConfig from "./ui.data-list/config.ts";
 import USelectConfig from "./ui.form-select/config.ts";
 import UListboxConfig from "./ui.form-listbox/config.ts";
@@ -259,12 +259,12 @@ export interface Components {
   ULabel: Partial<typeof ULabelConfig>;
   UColorPicker: Partial<typeof UColorPickerConfig>;
   UInput: Partial<typeof UInputConfig>;
-  UInputNumber: Partial<typeof UInputNumberConfig>;
+  UInputCounter: Partial<typeof UInputCounterConfig>;
   UInputPassword: Partial<typeof UInputPasswordConfig>;
   UInputRating: Partial<typeof UInputRatingConfig>;
   UInputSearch: Partial<typeof UInputSearchConfig>;
   UInputFile: Partial<typeof UInputFileConfig>;
-  UInputMoney: Partial<typeof UInputMoneyConfig>;
+  UInputNumber: Partial<typeof UInputNumberConfig>;
   UDataList: Partial<typeof UDataListConfig>;
   USelect: Partial<typeof USelectConfig>;
   [key: string]: UnknownObject;
