--- src/ui.form-input-money/config.ts
@@ -1,17 +1,0 @@
-export default /*tw*/ {
-  moneyInput: "{UInput}",
-  defaults: {
-    symbol: "",
-    size: "md",
-    minFractionDigits: 0,
-    maxFractionDigits: 2,
-    decimalSeparator: ",",
-    thousandsSeparator: " ",
-    labelAlign: "topInside",
-    prefix: "",
-    valueType: "number",
-    positiveOnly: false,
-    readonly: false,
-    disabled: false,
-  },
-};
