+++ src/ui.form-input-number/config.ts
@@ -1,54 +1,17 @@
 export default /*tw*/ {
-  label: {
-    base: "{ULabel}",
-    content: {
-      base: "items-center",
-      variants: {
-        size: {
-          sm: "gap-3",
-          md: "gap-3.5",
-          lg: "gap-4",
-        },
-      },
-    },
-  },
-  number: "",
-  addButton: "{UButton}",
-  removeButton: "{UButton}",
-  actionIcon: {
-    base: "{UIcon}",
-    defaults: {
-      size: {
-        sm: "xs",
-        md: "sm",
-        lg: "md",
-      },
-    },
-  },
-  addIcon: "{>actionIcon}",
-  removeIcon: "{>actionIcon}",
-  value: {
-    base: "font-bold select-none !leading-none",
-    variants: {
-      size: {
-        sm: "text-small",
-        md: "text-medium",
-        lg: "text-large",
-      },
-      disabled: {
-        true: "pointer-events-none",
-      },
-    },
-  },
+  numberInput: "{UInput}",
   defaults: {
+    currency: "",
     size: "md",
-    labelAlign: "topWithDesc",
+    minFractionDigits: 0,
+    maxFractionDigits: 2,
+    decimalSeparator: ",",
+    thousandsSeparator: " ",
+    labelAlign: "topInside",
+    prefix: "",
+    valueType: "number",
+    positiveOnly: false,
+    readonly: false,
     disabled: false,
-    step: 1,
-    min: 1,
-    max: 999,
-    /* icons */
-    removeIcon: "remove",
-    addIcon: "add",
   },
 };
