+++ src/ui.form-input-number/constants.ts
@@ -3,3 +3,5 @@
  `defineProps` is referencing locally declared variables. (vue/valid-define-props)
  */
 export const COMPONENT_NAME = "UInputNumber";
+
+export const RAW_DECIMAL_MARK = ".";
