+++ src/ui.form-input-number/storybook/stories.ts
@@ -5,28 +5,38 @@ import {
   getDocsDescription,
 } from "../../utils/storybook.ts";
 
-import UInputNumber from "../../ui.form-input-number/UInputNumber.vue";
+import UInputNumber from "../UInputNumber.vue";
 import UCol from "../../ui.container-col/UCol.vue";
-import UBadge from "../../ui.text-badge/UBadge.vue";
+import UIcon from "../../ui.image-icon/UIcon.vue";
+import UButton from "../../ui.button/UButton.vue";
+import URow from "../../ui.container-row/URow.vue";
+import UAvatar from "../../ui.image-avatar/UAvatar.vue";
 
 import type { Meta, StoryFn } from "@storybook/vue3";
 import type { Props } from "../types.ts";
 
 interface UInputNumberArgs extends Props {
   slotTemplate?: string;
-  enum: "size" | "labelAlign";
+  enum: "labelAlign" | "size";
 }
 
+const argTypes = getArgTypes(UInputNumber.__name);
+
 export default {
-  id: "3040",
+  id: "3030",
   title: "Form Inputs & Controls / Input Number",
   component: UInputNumber,
   args: {
-    modelValue: 1,
-    label: "Choose the number of items",
+    label: "Expected amount",
+    modelValue: 245000.42,
   },
   argTypes: {
-    ...getArgTypes(UInputNumber.__name),
+    ...argTypes,
+    valueType: {
+      ...argTypes?.valueType,
+      options: (argTypes?.valueType?.table?.type?.summary as string)?.split(" | "),
+      control: "select",
+    },
   },
   parameters: {
     docs: {
@@ -36,14 +46,18 @@ export default {
 } as Meta;
 
 const DefaultTemplate: StoryFn<UInputNumberArgs> = (args: UInputNumberArgs) => ({
-  components: { UInputNumber, UBadge },
+  components: { UInputNumber, UIcon, UButton },
   setup() {
     const slots = getSlotNames(UInputNumber.__name);
 
     return { args, slots };
   },
   template: `
-    <UInputNumber v-bind="args" v-model="args.modelValue">
+    <UInputNumber
+      v-bind="args"
+      v-model="args.modelValue"
+      class="max-w-96"
+    >
       ${args.slotTemplate || getSlotsFragment("")}
     </UInputNumber>
   `,
@@ -52,89 +66,159 @@ const DefaultTemplate: StoryFn<UInputNumberArgs> = (args: UInputNumberArgs) => (
 const EnumVariantTemplate: StoryFn<UInputNumberArgs> = (args: UInputNumberArgs, { argTypes }) => ({
   components: { UInputNumber, UCol },
   setup() {
+    let filteredOptions = argTypes?.[args.enum]?.options;
+
+    if (args.enum === "labelAlign") {
+      filteredOptions = argTypes?.[args.enum]?.options?.filter(
+        (item) => item !== "right" && item !== "topWithDesc",
+      );
+    }
+
     return {
       args,
-      options: argTypes?.[args.enum]?.options,
+      filteredOptions,
     };
   },
   template: `
-    <UCol gap="xl">
+    <UCol>
       <UInputNumber
-        v-for="(option, index) in options"
+        v-for="(option, index) in filteredOptions"
         :key="index"
-        v-bind="args"
-        :label="option"
         :[args.enum]="option"
+        label="Expected amount"
+        :placeholder="option"
+        class="max-w-96"
       />
     </UCol>
   `,
 });
 
 export const Default = DefaultTemplate.bind({});
-Default.args = { step: 1, min: 1, max: 100 };
+Default.args = {};
+
+export const Placeholder = DefaultTemplate.bind({});
+Placeholder.args = { placeholder: "Enter amount in USD" };
 
 export const Description = DefaultTemplate.bind({});
-Description.args = {
-  label: "Choose your storage capacity (in GB)",
-  description: "Storage capacity may not exceed 40 GB.",
-  max: 40,
-};
+Description.args = { description: "Please enter the transaction amount." };
 
 export const Error = DefaultTemplate.bind({});
-Error.args = { modelValue: NaN, error: "Passed value has incorrect format." };
+Error.args = {
+  modelValue: -245000.42,
+  error: "Invalid amount. Please enter a positive number with up to two decimal places.",
+};
+
+export const ReadOnly = DefaultTemplate.bind({});
+ReadOnly.args = { readonly: true };
 
 export const Disabled = DefaultTemplate.bind({});
 Disabled.args = { disabled: true };
 
+export const Currency = DefaultTemplate.bind({});
+Currency.args = { currency: "€" };
+
 export const Sizes = EnumVariantTemplate.bind({});
-Sizes.args = {
-  enum: "size",
-  modelValue: 1,
-  step: 1,
-  min: 1,
-  max: 100,
-};
+Sizes.args = { enum: "size" };
 
-export const LabelPlacement = EnumVariantTemplate.bind({});
-LabelPlacement.args = { enum: "labelAlign", description: "You may configure step value." };
+export const LabelAlign = EnumVariantTemplate.bind({});
+LabelAlign.args = { enum: "labelAlign" };
 
-export const ValueLimit = DefaultTemplate.bind({});
-ValueLimit.args = {
-  modelValue: 7,
-  step: 1,
-  min: 5,
-  max: 10,
-  label: "Min is 5 | Max is 10",
+export const LimitFractionDigits = DefaultTemplate.bind({});
+LimitFractionDigits.args = {
+  minFractionDigits: 4,
+  maxFractionDigits: 6,
+  description: "You can enter from 4 to 6 decimal places.",
 };
-ValueLimit.parameters = {
+LimitFractionDigits.parameters = {
   docs: {
     description: {
-      story: "To set the minimum and maximum values, use the `min` and `max` props.",
+      story:
+        // eslint-disable-next-line vue/max-len
+        "`minFractionDigits` and `maxFractionDigits` props determine the minimum/maximum number of digits to display after the decimal separator.",
     },
   },
 };
 
-export const Step = DefaultTemplate.bind({});
-Step.args = {
-  modelValue: 1,
-  step: 5,
-  min: 1,
-  max: 100,
-  label: "Step is 5",
-};
-Step.parameters = {
+export const DecimalSeparator = DefaultTemplate.bind({});
+DecimalSeparator.args = { decimalSeparator: "." };
+DecimalSeparator.parameters = {
   docs: {
     description: {
-      story: "`step` prop determines the increment/decrement value.",
+      story: "A symbol used to separate the integer part from the fractional part of a number.",
     },
   },
 };
 
-export const SlotLabel = DefaultTemplate.bind({});
-SlotLabel.args = {
-  slotTemplate: `
-    <template #label="{ label }">
-      <UBadge :label="label" />
-    </template>
+export const ThousandsSeparator = DefaultTemplate.bind({});
+ThousandsSeparator.args = { thousandsSeparator: "-" };
+ThousandsSeparator.parameters = {
+  docs: {
+    description: {
+      story: "A symbol used to separate the thousand parts of a number.",
+    },
+  },
+};
+
+export const PositiveOnly = DefaultTemplate.bind({});
+PositiveOnly.args = { positiveOnly: true };
+PositiveOnly.parameters = {
+  docs: {
+    description: {
+      story: "Allow only positive values.",
+    },
+  },
+};
+
+export const Prefix = DefaultTemplate.bind({});
+Prefix.args = { prefix: "+" };
+Prefix.parameters = {
+  docs: {
+    description: {
+      story: "Prefix to display before input value.",
+    },
+  },
+};
+
+export const IconProps: StoryFn<UInputNumberArgs> = (args) => ({
+  components: { UInputNumber, URow },
+  setup() {
+    return { args };
+  },
+  template: `
+    <URow>
+      <UInputNumber
+        left-icon="payments"
+        label="Annual payment"
+        placeholder="Enter your annual payment"
+      />
+      <UInputNumber
+        right-icon="currency_exchange"
+        label="Total amount"
+        currency="£"
+        placeholder="Enter the amount you want to exchange"
+      />
+    </URow>
   `,
-};
+});
+
+export const Slots: StoryFn<UInputNumberArgs> = (args) => ({
+  components: { UInputNumber, URow, UButton, UAvatar },
+  setup() {
+    return { args };
+  },
+  template: `
+    <URow>
+      <UInputNumber v-bind="args">
+        <template #left>
+          <UAvatar />
+        </template>
+      </UInputNumber>
+
+      <UInputNumber v-bind="args" :config="{ moneyInput: { rightSlot: 'pr-0' } }">
+        <template #right>
+          <UButton label="Calculate" size="sm" class="rounded-l-none h-full" />
+        </template>
+      </UInputNumber>
+    </URow>
+  `,
+});
