+++ src/ui.form-input-number/types.ts
@@ -4,36 +4,45 @@ import type { ComponentConfig } from "../types.ts";
 
 export type Config = typeof defaultConfig;
 
+export interface FormatOptions {
+  minFractionDigits: number;
+  maxFractionDigits: number;
+  decimalSeparator: string;
+  thousandsSeparator: string;
+  positiveOnly: boolean;
+  prefix: string;
+}
+
 export interface Props {
   /**
    * Input value.
    */
-  modelValue: number;
+  modelValue?: number | string | null;
 
   /**
-   * Input step.
+   * Set type of the modelValue.
    */
-  step?: number;
+  valueType?: "number" | "string";
 
   /**
-   * Input min value.
-   */
-  min?: number;
-
-  /**
-   * Input max value.
-   */
-  max?: number;
-
-  /**
-   * Input label below number.
+   * Input label.
    */
   label?: string;
 
   /**
    * Label placement.
    */
-  labelAlign?: "top" | "topWithDesc" | "left" | "right";
+  labelAlign?: "top" | "topInside" | "topWithDesc" | "left" | "right";
+
+  /**
+   * Currency symbol.
+   */
+  currency?: string;
+
+  /**
+   * Input placeholder.
+   */
+  placeholder?: string;
 
   /**
    * Input description.
@@ -51,9 +60,59 @@ export interface Props {
   size?: "sm" | "md" | "lg";
 
   /**
+   * Left icon name.
+   */
+  leftIcon?: string;
+
+  /**
+   * Right icon name.
+   */
+  rightIcon?: string;
+
+  /**
+   * Minimal number of signs after the decimal separator (fractional part of a number).
+   */
+  minFractionDigits?: number;
+
+  /**
+   * Maximal number of signs after the decimal separator (fractional part of a number).
+   */
+  maxFractionDigits?: number;
+
+  /**
+   * A symbol used to separate the integer part from the fractional part of a number.
+   */
+  decimalSeparator?: string;
+
+  /**
+   *  A symbol used to separate the thousand parts of a number.
+   */
+  thousandsSeparator?: string;
+
+  /**
+   * Allow only positive values.
+   */
+  positiveOnly?: boolean;
+
+  /**
+   * Prefix to display before input value.
+   */
+  prefix?: string;
+
+  /**
+   * Set input read-only.
+   */
+  readonly?: boolean;
+
+  /**
    * Disable the input.
    */
   disabled?: boolean;
+
+  /**
+   * Unique element id.
+   */
+  id?: string;
 
   /**
    * Component config object.
