+++ src/ui.form-input-number/utilFormat.ts
